call clean.sh

+++ scripts/build.sh
@@ -21,12 +21,11 @@ if [ -z "$VIRTUAL_ENV" ]; then
     exit 1
 fi
 
-echo "Removing 'build/...'..."
-rm -rf "build"
-
-# Uncomment the following lines if you want to remove 'dist/' as well
-# echo "Removing 'dist/...'..."
-# rm -rf "dist"
+bash scripts/clean.sh
+if [ $? -ne 0 ]; then
+    echo "Failed to clean project"
+    exit 1
+fi
 
 # Build exe
 rm -f "dist/osu-performance-analyzer" 2>/dev/null
